backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pandas==2.2.0
pydantic==2.6.3
pyshp==3.0.3
python-dotenv==1.0.1
networkx>=3.0
pyyaml>=6.0
geopandas==1.1.3
fiona==1.10.1
shapely==2.1.2
pyproj==3.7.1
pyogrio==0.12.1
